ai_evaluator/requirements.txt
openai
langchain
langchain-text-splitters
langchain-openai
langchain-community
faiss-cpu
pymupdf
python-pptx
playwright
tqdm
sentence-transformers